Add ANN by Euclidean distance using scalar-random-projection LSH
This adds support for ANN for Euclidean distances using
scalar-random-projection LSH (AKA locality-sensitive hashing based on
p-stable distributions).

As a slight optimization, it uses a random projection matrix to
represent the collection of random "a" vectors in the hash functions,
and computes their dot products with the input vector using
matrix-vector multiplication.