Repartition candidates after identifying them from hash tables
With the points included in the hash tables, the size of the tables is
now larger, which increases the memory pressure on executors. Since the
density of points is unlikely to be uniform and the LSH functions
involve some randomness, the identified candidate pairs will suffer from
partition skew. That makes it important to repartition after identifying
candidate pairs, both to reduce the memory pressure and to speed up
subsequent steps (e.g. distance computation).