Persist the hash tables in memory/on disk
After including the points in the hash tables, less joins are required,
which makes persistence less of an issue. The candidate pairs, for
example, are no longer persisted, because there's no join and therefore
no need to avoid recomputing the RDD for each half of the join.

This change removes persistence from CandidateStrategy and (mostly) from
ANNModel. The hash tables are now persisted while building the model,
which allows the removal of persistenceLevel as a model member.