Revert sbt-spark-package commits
It turns out that sbt-spark-package works with very specific settings for
registering a package to the index, but that those settings do not produce a
resolvable package when using `sbt spPublishLocal` (and also break
`sbt package publishLocal`). It's not clear if Databricks is actively supporting
or maintaining the plugin, so it's probably best to remove from this project.