Use `cogroup` instead of `join` to build candidate pairs
Using `join` on the hash tables to build the list of candidate pairs results in
many, many copies of the point vectors (one for each candidate neighbor). The
copies take up a lot of memory, and avoiding materializing all of the candidate
pairs would save a lot of memory.

An RDD `join` consists of a `cogroup` followed by a `flatMap`. The output of
`cogroup` is `(K, Iterable[V], Iterable[W])`, and when an RDD is cogrouped with
itself, the output will contain only two copies of each vector. That would be a
big improvement on a `join`, if the output of a `cogroup` can be used directly.

In order to do that, this change combines the distance computation with the
general structure of the `flatMap` used in a `join`. This avoids materializing
vector pairs and outputs `((id1, id2), distance)` records, which are much
smaller.